pyproject.toml
[project]
name = "exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "black>=26.3.1",
    "datasets>=4.8.4",
    "evaluate>=0.4.6",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mypy>=1.20.0",
    "numpy>=2.4.4",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "plotly>=5.24",
    "polars>=1.39.3",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.1",
]
